Fix: android build and transition for bars
