--- android/app/src/main/kotlin/com/example/word_by_word_game/MainActivity.kt
@@ -1,6 +1,0 @@
-package dev.xsoulspace.word_by_word_game
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}
